requirements.txt
pandas
numpy
matplotlib
scikit-learn
pyyaml
requests
imbalanced-learn
joblib
requests